(Refactor) 福祉/育児のディレクトリにファイル分割する
